boutique_pro/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
Pillow>=10.0

